Use 64-bit floats for call tree timings.
Fixes #5310.

This matches the precision of regular JavaScript numbers.
32-bit floats don't have enough precision to exactly represent
integers larger than 16,777,216. So they don't have sufficient
precision for the call tree which sometimes is used to with
bytes values which can range in the gigabytes.

+++ src/types/profile-derived.js
@@ -801,7 +801,7 @@ export type SortedTabPageData = Array<{|
 export type CallNodeSelfAndSummary = {|
   // This property stores the amount of unit (time, bytes, count, etc.) spent in
   // this call node and not in any of its descendant nodes.
-  callNodeSelf: Float32Array,
+  callNodeSelf: Float64Array,
   // The sum of absolute values in callNodeSelf.
   // This is used for computing the percentages displayed in the call tree.
   rootTotalSummary: number,
